@@ -3,12 +3,27 @@
   "version": "0.0.0",
   "description": "Description for mica",
   "dependencies": {
-    "admin-lte": "^3.2.0",
+    "admin-lte": "^4.0.0-rc4",
+    "bootstrap": "^5.3.3",
+    "@popperjs/core": "^2.11.8",
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "jquery": "^3.7.1",
     "axios": "^1.13.5",
     "bootstrap-4-autocomplete": "^1.3.2",
     "bootstrap-datepicker": "^1.9.0",
-    "chart.js": "^2.9.4",
-    "chartjs-chart-geo": "^2.1.0",
+    "chart.js": "^4.4.6",
+    "chartjs-chart-geo": "^4.3.3",
+    "datatables.net": "^1.13.11",
+    "datatables.net-bs5": "^1.13.11",
+    "datatables.net-buttons": "^2.4.3",
+    "datatables.net-buttons-bs5": "^2.4.3",
+    "datatables.net-fixedheader": "^3.4.0",
+    "datatables.net-fixedheader-bs5": "^3.4.0",
+    "datatables.net-select": "^1.7.0",
+    "datatables.net-select-bs5": "^1.7.0",
+    "select2": "^4.1.0-rc.0",
+    "select2-bootstrap-5-theme": "^1.3.0",
+    "toastr": "^2.1.4",
     "jquery.redirect": "^1.1.4",
     "js-cookie": "^2.2.1",
     "marked": "^15.0.6",
@@ -17,9 +32,16 @@
     "plotly.js-dist-min": "^2.5.1",
     "rql": "github:epigeny/rql#1.1.0",
     "simplemde": "^1.11.2",
-    "vue": "^2.6.11"
+    "vue": "^3.4.0",
+    "@vue/compat": "^3.4.0"
+  },
+  "scripts": {
+    "build:css": "sass --quiet-deps --style=compressed --source-map --load-path=node_modules src/main/resources/scss/mica.scss src/main/webapp/assets/css/mica.min.css",
+    "build:css:prod": "sass --quiet-deps --style=compressed --no-source-map --load-path=node_modules src/main/resources/scss/mica.scss src/main/webapp/assets/css/mica.min.css",
+    "watch:css": "sass --quiet-deps --watch --source-map --load-path=node_modules src/main/resources/scss/mica.scss src/main/webapp/assets/css/mica.min.css"
   },
   "devDependencies": {
+    "sass": "^1.93.2",
     "grunt": "1.0.1",
     "grunt-autoprefixer": "3.0.4",
     "grunt-cli": "1.2.0",
