@@ -1,61 +1,4 @@
-class StringLocalizer {
-  static __localizeInternal(entries, locale) {
-    const result = (Array.isArray(entries) ? entries : [entries]).filter((entry) => entry && (locale === entry.lang || locale === entry.locale)).pop();
-
-    if (result) {
-      let value = result.value ? result.value : result.text;
-      return value ? value : null;
-    }
-    return null;
-  }
-
-  static localize(entries) {
-    if (entries) {
-      const result = StringLocalizer.__localizeInternal(entries, Mica.locale)
-        || StringLocalizer.__localizeInternal(entries, Mica.defaultLocale)
-        || StringLocalizer.__localizeInternal(entries, 'und');
-
-      return result ? result : '';
-    } else {
-      return '';
-    }
-  }
-}
-
-// Register all filters
-
-Vue.filter("ellipsis", (input, n, link) => {
-  if (input.length <= n) { return input; }
-  const subString = input.substr(0, n-1); // the original check
-  const anchor = link ? ` <a href="${link}">...</a>` : " ...";
-  return subString.substr(0, subString.lastIndexOf(" ")) + anchor;
-});
-
-Vue.filter("readmore", (input, link, text) => {
-  return `${input} <a href="${link}" class="clearfix btn-link">${text}</a>`;
-});
-
-Vue.filter("concat", (input, suffix) => {
-  return input + suffix;
-});
-
-Vue.filter("localize-string", (input) => {
-  if (typeof input === "string") return input;
-  return StringLocalizer.localize(input);
-});
-
-Vue.filter("markdown", (input) => {
-  return marked.parse(input);
-});
-
-Vue.filter("localize-number", (input) => {
-  return (input || 0).toLocaleString();
-});
-
-Vue.filter("translate", (key) => {
-  let value = Mica.tr[key];
-  return typeof value === "string" ? value : key;
-});
+// Filter functions are now provided by mica-filters.js (MicaFilters)
 
 /**
  * Base class for all entities (Srtudies, Networks, Datasets)
@@ -371,8 +314,8 @@ const StatItemComponent = {
     url: String
   },
   template: `
-    <a v-if="count" v-bind:href="url" class="btn btn-sm btn-link col text-left">
-      <span class="h6 pb-0 mb-0 d-block">{{count | localize-number}}</span>
+    <a v-if="count" v-bind:href="url" class="btn btn-sm btn-link col text-start">
+      <span class="h6 pb-0 mb-0 d-block">{{localizeNumber(count)}}</span>
       <span class="text-muted"><small>{{count < 2 ? this.singular : this.plural}}</small></span>
     </a>
   `
@@ -394,8 +337,8 @@ const VariableStatItemComponent =  {
     }
   },
   template: `
-    <a v-if="count" v-bind:href="url" class="btn btn-sm btn-link col text-left">
-      <span class="h6 pb-0 mb-0 d-block">{{count | localize-number}}</span>
+    <a v-if="count" v-bind:href="url" class="btn btn-sm btn-link col text-start">
+      <span class="h6 pb-0 mb-0 d-block">{{localizeNumber(count)}}</span>
       <span class="text-muted"><small>{{this.countLabel}}</small></span>
     </a>
   `,
@@ -429,8 +372,8 @@ const DatasetStatItemComponent =  {
     }
   },
   template: `
-    <a v-if="count" href="javascript:void(0)" style="cursor: initial;" class="btn btn-sm col text-left">
-      <span class="h6 pb-0 mb-0 d-block">{{count | localize-number}}</span>
+    <a v-if="count" href="javascript:void(0)" style="cursor: initial;" class="btn btn-sm col text-start">
+      <span class="h6 pb-0 mb-0 d-block">{{localizeNumber(count)}}</span>
       <span class="text-muted"><small>{{this.countLabel}}</small></span>
     </a>
   `,
@@ -458,8 +401,8 @@ const StudyStatItemComponent =  {
     }
   },
   template: `
-    <a v-if="count" href="javascript:void(0)" style="cursor: initial;" class="btn btn-sm col text-left">
-      <span class="h6 pb-0 mb-0 d-block">{{count | localize-number}}</span>
+    <a v-if="count" href="javascript:void(0)" style="cursor: initial;" class="btn btn-sm col text-start">
+      <span class="h6 pb-0 mb-0 d-block">{{localizeNumber(count)}}</span>
       <span class="text-muted"><small>{{this.countLabel}}</small></span>
     </a>
   `,
@@ -486,9 +429,9 @@ const StudyStatItemComponent =  {
  */
 const EntitiesSortingComponent = {
   template: `
-    <div class="sorting position-relative float-left">
+    <div class="sorting position-relative float-start">
       <div class="dropdown">
-        <button type="button" class="btn btn-outline-primary dropdown-toggle" data-toggle="dropdown">
+        <button type="button" class="btn btn-outline-primary dropdown-toggle" data-bs-toggle="dropdown">
           <span v-html="sortLabel(selectedChoice)"></span>
         </button>
 
@@ -540,17 +483,18 @@ const EntitiesSortingComponent = {
  * Typeahead UI component
  */
 const TypeaheadComponent = {
+  emits: ['typing', 'select'],
   template: `
     <div class="typeahead w-100 position-relative">
       <div class="input-group">
-        <input type="text" :placeholder="'listing-typeahead-placeholder'   | translate" class="form-control form-control-sm" v-model="text" @keyup="typing($event)">
-        <div class="input-group-append">
-          <button type="button" class="btn btn-primary btn-sm" @click="select(text)"><i class="fas fa-filter"></i></button>
+        <div class="position-relative flex-grow-1">
+          <input type="text" :placeholder="translate('listing-typeahead-placeholder')" class="form-control form-control-sm" :class="{'pe-4': text.length > 0}" v-model="text" @keyup="typing($event)" @input="onInput">
+          <button type="button" class="btn btn-link btn-sm text-muted position-absolute top-50 end-0 translate-middle-y px-2 py-0" v-show="text.length > 0" @click="clear" style="z-index:5;"><i class="fas fa-times"></i></button>
         </div>
-        <button type="button" class="close position-absolute" style="right: 2em; top: 0.25em;" v-if="text.length > 0" @click="clear"><span aria-hidden="true">&times;</span></button>
+        <button type="button" class="btn btn-primary btn-sm" @click="select(text)"><i class="fas fa-filter"></i></button>
       </div>
 
-      <div class="list-group position-absolute mt-1 ml-1 shadow" style="z-index: 100; overflow-y: auto; max-height: 16em;" v-if="showChoices && typeaheadItems.length > 0">
+      <div class="list-group position-absolute mt-1 ms-1 shadow" style="z-index: 100; overflow-y: auto; max-height: 16em;" v-if="showChoices && typeaheadItems.length > 0">
         <button type="button" class="list-group-item list-group-item-action" :class="{ active: index === currentIndexSelection }" v-for="(item, index) in typeaheadItems" :key="item" v-html="highlight(item)" @click="select(quote(item))"></button>
       </div>
     </div>
@@ -568,7 +512,7 @@ const TypeaheadComponent = {
   },
   data() {
     return {
-      text: this.externalText || "",
+      text: String(this.externalText || ""),
       showChoices: false,
       currentIndexSelection: -1,
     };
@@ -579,19 +523,24 @@ const TypeaheadComponent = {
     },
   },
   methods: {
+    onInput() {
+      if (this.text.length === 0) {
+        this.$emit('select', '');
+      }
+    },
     typing(event) {
-      if (event.keyCode === 13) {
+      if (event.key === 'Enter') {
         this.select(this.text);
         return;
       }
 
-      if (event.keyCode === 27) { // escape
+      if (event.key === 'Escape') { // escape
         this.clear();
         return;
       }
 
-      if (event.keyCode === 37 || event.keyCode === 38 || event.keyCode === 39 || event.keyCode === 40) { // arrows
-        this.changeIndexSelection(event.keyCode);
+      if (['ArrowLeft', 'ArrowUp', 'ArrowRight', 'ArrowDown'].includes(event.key)) { // arrows
+        this.changeIndexSelection(event.key);
         return;
       }
 
@@ -610,14 +559,14 @@ const TypeaheadComponent = {
     clear() {
       this.select('');
     },
-    changeIndexSelection(keyCode) {
-      if (keyCode === 37 || keyCode === 38) { // up
+    changeIndexSelection(key) {
+      if (key === 'ArrowLeft' || key === 'ArrowUp') { // up
         if (this.currentIndexSelection <= 0) {
           this.currentIndexSelection = this.typeaheadItems.length;
         }
 
         this.currentIndexSelection--;
-      } else if (keyCode === 39 || keyCode === 40) { // down
+      } else if (key === 'ArrowRight' || key === 'ArrowDown') { // down
         if (this.currentIndexSelection >= this.typeaheadItems.length - 1) {
           this.currentIndexSelection = -1;
         }
@@ -642,7 +591,7 @@ const TypeaheadComponent = {
       return text;
     },
     cleanUnclosedDoubleQuotes(text) {
-      let output = (text || "").trim();
+      let output = (String(text == null ? '' : text)).trim();
       const doubleQuotesRegxp = /"/g;
       const instancesOfDoubleQuoteCharacters = (output.match(doubleQuotesRegxp) || []).length;
 
@@ -661,7 +610,7 @@ const TypeaheadComponent = {
  */
 const NAVIGATION_POSITIONS = ['bottom', 'top'];
 
-const ObibaEntitiesApp = {
+const ObibaEntitiesAppMixin = {
   data() {
     return {
       loading: false,
@@ -783,12 +732,11 @@ const ObibaEntitiesApp = {
 class ObibaDatasetsApp {
 
   static build(element, type, locale, sortOptionsTranslations) {
-    return new Vue({
-      locale,
-      el: element,
-      extends: ObibaEntitiesApp,
+    const app = MicaVueApp.createApp({
+      mixins: [ObibaEntitiesAppMixin],
       data() {
         return {
+          locale,
           sortOptionsTranslations,
           initialSort: 'name'
         };
@@ -846,15 +794,15 @@ class ObibaDatasetsApp {
         }
       }
     });
+    return app.mount(element);
   }
 }
 
 class ObibaStudiesApp {
 
   static build(element, type, locale, sortOptionsTranslations, initialSort) {
-    return new Vue({
-      el: element,
-      extends: ObibaEntitiesApp,
+    const app = MicaVueApp.createApp({
+      mixins: [ObibaEntitiesAppMixin],
       data() {
         return {
           locale,
@@ -905,15 +853,15 @@ class ObibaStudiesApp {
         }
       }
     });
+    return app.mount(element);
   }
 }
 
 class ObibaNetworksApp {
 
   static build(element, locale, sortOptionsTranslations) {
-    return new Vue({
-      el: element,
-      extends: ObibaEntitiesApp,
+    const app = MicaVueApp.createApp({
+      mixins: [ObibaEntitiesAppMixin],
       data() {
         return {
           locale,
@@ -970,5 +918,6 @@ class ObibaNetworksApp {
         }
       }
     });
+    return app.mount(element);
   }
 }
